Add files via upload
I developed a ticketing system. Designed to optimize ticket sales management for events. To secure the data from unauthorized access, users must enter a password before accessing the system. The data recorded includes the name of the individual, the date of the record, the fine amount, ticket type, event name, number of tickets purchased, and total price. The entered data is stored systematically and can be accessed at any time, facilitating the tracking and analysis of ticket sales.